+++ Server/KFPlugin/KFRouteProxy/KFRouteZoneManage.h
@@ -16,8 +16,8 @@ namespace KFrame
         }
 
         uint32 _zone_id;
-        uint32 _server_id;
-        uint32 _handle_id;
+        uint64 _server_id;
+        uint64 _handle_id;
     };
 
     // 匹配服务器管理器
@@ -28,17 +28,17 @@ namespace KFrame
         ~KFRouteZoneManage() = default;
 
         // 添加游戏区
-        KFRouteZone* AddRouteZone( uint32 zoneid, uint32 serverid, uint32 handleid );
+        KFRouteZone* AddRouteZone( uint32 zoneid, uint64 serverid, uint64 handleid );
 
         // 删除游戏区
-        void RemoveRouteZone( uint32 serverid );
+        void RemoveRouteZone( uint64 serverid );
 
         // 查找登录节点
-        KFRouteZone* FindRouteZone( uint32 serverid );
+        KFRouteZone* FindRouteZone( uint64 serverid );
 
     public:
         // 登录节点列表
-        KFMap< uint32, uint32, KFRouteZone > _kf_route_zone;
+        KFMap< uint64, uint64, KFRouteZone > _kf_route_zone;
     };
 
     ///////////////////////////////////////////////////////////////////////////////////////////////
