+++ Server/KFPlugin/KFRouteShard/KFRouteProxyManage.cpp
@@ -12,7 +12,7 @@ namespace KFrame
 
     }
 
-    KFRouteProxy* KFRouteProxyManage::AddRouteProxy( uint32 zoneid, uint32 serverid, uint32 handleid )
+    KFRouteProxy* KFRouteProxyManage::AddRouteProxy( uint32 zoneid, uint64 serverid, uint64 handleid )
     {
         auto routezone = _kf_route_proxy.Create( serverid );
         routezone->_zone_id = zoneid;
@@ -22,12 +22,12 @@ namespace KFrame
         return routezone;
     }
 
-    void KFRouteProxyManage::RemoveRouteProxy( uint32 serverid )
+    void KFRouteProxyManage::RemoveRouteProxy( uint64 serverid )
     {
         _kf_route_proxy.Remove( serverid );
     }
 
-    KFRouteProxy* KFRouteProxyManage::FindRouteProxy( uint32 serverid )
+    KFRouteProxy* KFRouteProxyManage::FindRouteProxy( uint64 serverid )
     {
         return _kf_route_proxy.Find( serverid );
     }
