+++ Server/KFPlugin/KFZone/KFZoneModule.h
@@ -35,8 +35,8 @@ namespace KFrame
         virtual const KFZone* GetZone() const;
 
         // 判断游戏分区id
-        virtual bool IsServerSameZone( uint32 serverid );
-        virtual bool IsPlayerSameZone( uint32 playerid );
+        virtual bool IsServerSameZone( uint64 serverid );
+        virtual bool IsPlayerSameZone( uint64 playerid );
     };
 }
